requirements.txt
tweepy
python-dotenv
schedule
requests
openai